refactor:usePostDetail 분리
